# authz-authn/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lyfter JWT
  description: API specification generated from Flask blueprints for User, Product, Invoice, Order, and Contact services.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
paths:
...
  /contacts/:
    get:
      summary: List contacts
      description: Returns contacts for the authenticated user. Admins can pass `user_id` to inspect contacts belonging to a specific user.
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: user_id
          required: false
          schema:
            type: integer
          description: Filter contacts by user ID (admin privilege required to query other users)
      responses:
        '200':
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /contacts/{id}:
    get:
      summary: Get contact by ID
      description: Retrieves a specific contact. Non-admins can only retrieve contacts that belong to them.
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Unique contact ID
      responses:
        '200':
          description: Contact retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          description: Contact not found or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update a contact
      description: Updates name, phone number, and email. Non-admins can only update their own contacts.
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Unique contact ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactUpdateRequest'
      responses:
        '200':
          description: Contact updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          description: Invalid JSON or parameter validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contact not found or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a contact by ID
      description: Deletes a contact record. Non-admins can only delete their own contacts.
      tags:
        - Contacts
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Unique contact ID
      responses:
        '200':
          description: Contact deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          description: Contact not found or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========================================
  # PRODUCT ROUTES (/products)
  # ==========================================
...
  /invoices/:
    get:
      summary: Retrieve invoices
      description: Retrieve all invoices or filter by user ID using query parameters. Requires admin privileges.
      tags:
        - Invoices
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: user_id
          required: false
          schema:
            type: integer
          description: Filter invoices by a specific user ID
      responses:
        '200':
          description: List of invoices (filtered or unfiltered)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
        '404':
          description: Invoices not found for the specified criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /invoices/{id}:
    get:
      summary: Get invoice by ID
      description: Requires admin privileges.
      tags:
        - Invoices
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Unique invoice ID
      responses:
        '200':
          description: Invoice retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '500':
          description: Unable to retrieve invoice or internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete an invoice by ID
      description: Requires admin privileges.
      tags:
        - Invoices
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Unique invoice ID
      responses:
        '200':
          description: Invoice deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '500':
          description: Unable to delete invoice or internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========================================
  # ORDER ROUTES (/orders)
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ContactUpdateRequest:
      type: object
      required:
        - name
        - phone_number
        - email
      properties:
        name:
          type: string
          example: Jane Doe
        phone_number:
          type: string
          example: "+15559876543"
        email:
          type: string
          format: email
          example: jane.updated@example.com
    Contact:
      type: object
      properties:
        id:
          type: integer
          example: 1001
        user_id:
          type: integer
          example: 10
        name:
          type: string
          example: Jane Doe
        phone_number:
          type: string
          example: "+15551234567"
        email:
          type: string
          format: email
          example: jane.doe@example.com
...
    InvoiceDetail:
      type: object
      properties:
        id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 101
        quantity:
          type: integer
          example: 2
        unit_price:
          type: number
          format: float
          example: 29.99
    Invoice:
      type: object
      properties:
        id:
          type: integer
          example: 5001
        user_id:
          type: integer
          example: 1
        details:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceDetail'
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
...